KEYCLOAK-4743 Add support for HTTP Proxy in DefaultHttpClientFactory
We now support basic proxy configuration for out-bound calls
via the HttpClient provided by the HttpClientProvider.

To configure a proxy one needs to set the proxy-url property to the
"default" connectionsHttpClient in the standalone(-ha).xml file.
The `HttpClientBuilder` builder will derive the proxy configuration
from the provided proxy-url.
By default no proxy is configured.

The following jboss-cli command can be used to configure
the proxy-url.
```
/subsystem=keycloak-server/spi=connectionsHttpClient/provider=default:write-attribute(name=properties.proxy-url,value="http://www-proxy.tdlabs.local:3128")
```